Add Task Model, Fix Code Scan issue, revise and adjustment for Group

+++ src/main/java/com/nus/iss/tasktracker/dto/UserDTO.java
@@ -6,9 +6,9 @@ import lombok.Data;
 @Data
 public class UserDTO {
 
-    private long userId;
+    private Integer userId;
 
-    private long groupId;
+    private Integer groupId;
 
     private String email;
